+++ src/lib/debug_info/debug_info_view.js
@@ -56,6 +56,8 @@ ${translate('debugInfo.details.postParams')}:       ${page.postParams}
 ${translate('debugInfo.details.sendCredentials')}:  ${page.sendCredentials}
 ${translate('debugInfo.details.waitForNetworkIdle')}: ${page.waitForNetworkIdle}
 ${translate('debugInfo.details.waitForNetworkIdleTimeoutMs')}: ${page.waitForNetworkIdleTimeoutMs}
+${translate('debugInfo.details.hiddenTabIgnoreSelectors')}: ${page.hiddenTabIgnoreSelectors}
+${translate('debugInfo.details.hiddenTabUseTextSnapshotHash')}: ${page.hiddenTabUseTextSnapshotHash}
 ${translate('debugInfo.details.state')}:            ${page.state}
 ${translate('debugInfo.details.lastAutoscanTime')}: ${lastAutoscanTime}
 ${translate('debugInfo.details.oldScanTime')}:      ${oldScanTime}